// libs/sales/product-finder/data-access/src/lib/entities/error/product-finder-error.enum.ts
export enum ProductFinderErrorEnum {
  REQUIRE_AT_LEAST_ONE_PREFERENCE_TO_BE_SELECTED = 'Require at least one preference to be selected',
  REQUIRE_INTERNET_WHEN_TV_IS_SELECTED = 'Require internet when tv is selected',
  REQUIRE_AT_LEAST_ONE_INTERNET_USAGE_TO_BE_SELECTED = 'Require at least one internet usage to be selected',
  REQUIRE_AT_LEAST_ONE_MOBILE_USAGE_TO_BE_SELECTED_FOR_EACH_MOBILE_LINE = 'Require at least one mobile usage to be selected for each mobile line',
  REQUIRE_AT_LEAST_ONE_MOBILE_USAGE_TO_BE_SELECTED = 'Require at least one mobile usage to be selected',
  REQUIRE_AT_LEAST_ONE_TV_CHANNEL_TO_BE_SELECTED = 'Require at least one tv channel to be selected',
  REQUIRE_AT_LEAST_ONE_TV_EXPERIENCE_TO_BE_SELECTED = 'Require at least one tv experience to be selected',
}